requirements.txt
alembic==1.10.2
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
dominate==2.7.0
email-validator==1.3.1
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.8
Werkzeug==2.3.6
WTForms==3.0.1
zipp==3.15.0
